requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.2.0
Automat==22.10.0
Babel==2.12.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
cssselect==1.2.0
Django==4.2.2
django-cors-headers==4.1.0
django-crispy-forms==2.0
django-csv-export==0.0.3
django-csv-export-view==2.0.0
django-currencies==0.10.1
django-datatables-view==1.20.0
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-extensions==3.2.3
django-filter==23.2
django-model-utils==4.3.1
django-money==3.1.0
django-redis==5.3.0
django-rest-swagger==2.2.0
django-taggit==4.0.0
django-taggit-serializer==0.1.7
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-datatables==0.7.0
djangorestframework-simplejwt==5.2.2
drf-jwt==1.19.2
exceptiongroup==1.2.0
geographiclib==2.0
geopy==2.4.1
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonfield==2.1.1
lxml==5.1.0
MarkupSafe==2.1.3
numpy==1.25.1
openapi-codec==1.3.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.0.3
parsel==1.8.1
Pillow==9.5.0
playwright==1.35.0
pluggy==1.3.0
psycopg2==2.9.6
py-moneyed==2.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyee==9.0.4
PyJWT==2.7.0
pyOpenSSL==23.3.0
PySocks==1.7.1
pytest==7.4.4
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
queuelib==1.6.2
redis==4.5.5
requests==2.31.0
Scrapy==1.7.3
selenium==4.17.2
service-identity==23.1.0
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
tomli==2.0.1
trio==0.24.0
trio-websocket==0.11.1
Twisted==21.7.0
typing_extensions==4.9.0
tzdata==2023.3
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.0.3
user-agent==0.1.10
user-agents==2.2.0
w3lib==2.1.2
webencodings==0.5.1
wsproto==1.2.0
zope.interface==6.1
